Use enum class in logger-config as it is possible to serialize now.
[Bug/Feature]   N/A
[Cause]         To keep things clean and consistent
[Solution]      N/A
[Verification]  Build, run the junk tester

+++ libs/lxcpp/logger-config.cpp
@@ -39,8 +39,8 @@ void LoggerConfig::set(const logger::LogType type,
         }
     }
 
-    mType = static_cast<int>(type);
-    mLevel = static_cast<int>(level);
+    mType = type;
+    mLevel = level;
     mArg = arg;
 }
